Constants: Move out of class
The Constants class was pretty useless. So, moved all the constants
outside the class. Now the class doesn't exist any more.
Made necessary changes to all the references.
Removed unnecessary comment in Constants.

Fixes https://github.com/coala-analyzer/coala/issues/917
