+++ coalib/tests/output/dbus/BuildDbusServiceTest.py
@@ -7,7 +7,7 @@ from distutils.errors import DistutilsOptionError
 
 sys.path.insert(0, ".")
 from coalib.output.dbus.BuildDbusService import BuildDbusService
-from coalib.misc.Constants import Constants
+from coalib.misc import Constants
 
 
 class BuildDbusServiceTest(unittest.TestCase):
