+++ coalib/tests/output/printers/LogPrinterTest.py
@@ -5,7 +5,7 @@ from pyprint.Printer import Printer
 from pyprint.NullPrinter import NullPrinter
 
 sys.path.insert(0, ".")
-from coalib.misc.Constants import Constants
+from coalib.misc import Constants
 from coalib.processes.communication.LogMessage import LogMessage, LOG_LEVEL
 from coalib.output.printers.LogPrinter import LogPrinter
 from coalib.output.printers.StringPrinter import StringPrinter
