+++ coalib/tests/output/printers/StringPrinterTest.py
@@ -2,7 +2,7 @@ import sys
 import unittest
 
 sys.path.insert(0, ".")
-from coalib.misc.Constants import Constants
+from coalib.misc import Constants
 from coalib.output.printers.StringPrinter import StringPrinter
