+++ coalib/tests/settings/ConfigurationGatheringTest.py
@@ -8,7 +8,7 @@ from pyprint.NullPrinter import NullPrinter
 from pyprint.ClosableObject import close_objects
 
 sys.path.insert(0, ".")
-from coalib.misc.Constants import Constants
+from coalib.misc import Constants
 from coalib.parsing.StringProcessing import escape
 from coalib.settings.ConfigurationGathering import (gather_configuration,
                                                     find_user_config)
