+++ coalib/tests/settings/SectionTest.py
@@ -4,7 +4,7 @@ import sys
 sys.path.insert(0, ".")
 
 from coalib.settings.Section import Section, Setting, append_to_sections
-from coalib.misc.Constants import Constants
+from coalib.misc import Constants
 
 
 class SectionTest(unittest.TestCase):
